requirements.txt
google-play-scraper
langchain
langchain-groq
langchain-huggingface
sentence-transformers
scikit-learn
pandas
python-dotenv
tabulate
matplotlib
seaborn
